Move the InfoRow component out of MultisigProposalDetails directory
