+++ src/components/Proposals/MultisigProposalDetails/TxDetails.tsx
@@ -1,4 +1,4 @@
-import { Box, Divider, Flex, Text, Tooltip } from '@chakra-ui/react';
+import { Box, Divider, Text } from '@chakra-ui/react';
 import { format } from 'date-fns';
 import { formatInTimeZone } from 'date-fns-tz';
 import { useTranslation } from 'react-i18next';
@@ -7,50 +7,7 @@ import { createAccountSubstring } from '../../../hooks/utils/useDisplayName';
 import { MultisigProposal } from '../../../types';
 import { DEFAULT_DATE_TIME_FORMAT } from '../../../utils/numberFormats';
 import ContentBox from '../../ui/containers/ContentBox';
-import DisplayTransaction from '../../ui/links/DisplayTransaction';
-
-function TransactionOrText({ txHash, value }: { txHash?: string | null; value?: string }) {
-  return txHash ? <DisplayTransaction txHash={txHash} /> : <Text>{value}</Text>;
-}
-
-export function InfoRow({
-  property,
-  value,
-  txHash,
-  tooltip,
-}: {
-  property: string;
-  value?: string;
-  txHash?: string | null;
-  tooltip?: string;
-}) {
-  return (
-    <Flex
-      marginTop={4}
-      justifyContent="space-between"
-    >
-      <Text
-        textStyle="text-base-sans-regular"
-        color="chocolate.200"
-      >
-        {property}
-      </Text>
-      {tooltip === undefined ? (
-        <TransactionOrText
-          txHash={txHash}
-          value={value}
-        />
-      ) : (
-        <Tooltip label={tooltip}>
-          <TransactionOrText
-            txHash={txHash}
-            value={value}
-          />
-        </Tooltip>
-      )}
-    </Flex>
-  );
-}
+import InfoRow from '../../ui/proposal/InfoRow';
 
 export function TxDetails({ proposal }: { proposal: MultisigProposal }) {
   const { t } = useTranslation('proposal');
